Created a modal & cartModal component separately
