+++ src/components/atoms/Modal/Modal.scss
@@ -1,0 +1,14 @@
+.modal {
+  position: relative;
+  width: auto;
+
+  &__overlay {
+    position: fixed;
+    z-index: 99;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.7);
+  }
+}
