+++ src/components/organisms/CartModal/CartModal.scss
@@ -1,57 +1,42 @@
 @import "../../../styles/variables";
 
-.cartmodal {
-  position: relative;
-  width: auto;
+.cartmodal__section {
+  width: 35vw;
+  height: 80vh;
+  max-width: 40rem;
+  font-size: $text-font-size;
+  background-color: $primary-color;
+  display: flex;
+  flex-direction: column;
+  position: fixed;
+  bottom: 0%;
+  left: 80%;
+  transform: translate(-50%, 0%);
+  z-index: 99;
 
-  &__overlay {
-    position: fixed;
-    z-index: 99;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.7);
+  &__flex {
+    flex: 1;
   }
 
-  &__section {
-    width: 35vw;
-    height: 80vh;
-    max-width: 40rem;
-    font-size: $text-font-size;
-    background-color: $primary-color;
+  &__heading {
     display: flex;
-    flex-direction: column;
-    position: fixed;
-    bottom: 0%;
-    left: 80%;
-    transform: translate(-50%, 0%);
-    z-index: 99;
+    background-color: #212121;
+    justify-content: space-between;
 
-    &__flex {
-      flex: 1;
+    &__title {
+      padding: 1rem 1rem;
+      font-size: $text-font-size;
+      color: $button-text-color;
     }
 
-    &__heading {
-      display: flex;
+    &__button {
       background-color: #212121;
-      justify-content: space-between;
+      border-width: 0;
+      margin-right: 1rem;
 
-      &__title {
-        padding: 1rem 1rem;
-        font-size: $text-font-size;
-        color: $button-text-color;
-      }
-
-      &__button {
-        background-color: #212121;
-        border-width: 0;
-        margin-right: 1rem;
-
-        &:active {
-          top: 0px;
-          left: 0px;
-        }
+      &:active {
+        top: 0px;
+        left: 0px;
       }
     }
   }
